Store on-disk caches in a temporary directory
As the cache system cannot support multiple caches running over the same
backend storage, we create a temporary cacheDir for each TeleportProcess
instance, and we use it to store all the sqlite backends for the caches.

With on-disk caches being single-use, we can run the underlying sqlite
backend with the MEMORY rollback journal to improve performance and
reduce filesystem activity.